feat(seo): adds OpenGraph, Twitter, and JSON-LD data for SEO
- Pages now support the following tags:
  - seo_title — shown in sharing, search results, browser tab
  - seo_description — shown in sharing, search results
  - seo_image — shown in sharing

These are all optional, and will default to sane values if
they’re omitted. I created an app on Facebook to avoid an
error in the debugger, but it won’t actually _do_ anything
right now.

I will also open an issue to remind myself about submitting
the site for Twitter card validation once this is merged in.

closes #18